selector cleanup
