+++ ui/pages/Search/components/filters/ProjectFamiliesFilter.jsx
@@ -6,11 +6,10 @@ import { Form, Button } from 'semantic-ui-react'
 import {
   getProjectsByGuid,
   getProjectAnalysisGroupFamilyGuidsByGuid,
-  getProjectDatasetTypes,
 } from 'redux/selectors'
 import { Multiselect, ButtonRadioGroup } from 'shared/components/form/Inputs'
 import { SOLVED_FAMILY_STATUS_OPTIONS } from 'shared/utils/constants'
-import { getProjectFamilies, getSearchContextIsLoading, getFamilyOptions, getAnalysisGroupOptions } from '../../selectors'
+import { getProjectFamilies, getSearchContextIsLoading, getFamilyOptions, getAnalysisGroupOptions, getProjectDatasetTypes } from '../../selectors'
 import { loadProjectFamiliesContext } from '../../reducers'
 import { ProjectFilter } from './ProjectsField'
